+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 -----
 - Add ``alleyoop-rates`` process
 - Add ``alleyoop-utr-rates`` process
+- Add ``alleyoop-summary`` process
 
 Changed
 -------